Trait logjam::ui::Ui

pub trait Ui {
    // Required method
    fn render(&self);

    // Provided method
    fn interacted(&self) -> bool { ... }
}

Trait for any structure related to rendering the User interface. Used in combination with the Menu struct to bundle multiple elements.

Required Methods

fn render(&self)

Displays element on screen. Required for any implementation

Provided Methods

fn interacted(&self) -> bool

Returns interaction with object. Used to handle actions with interactive elements. Ex. Buttons, sliders